port-range: fix handshake busy-loop and rr_counter overflow
From the branch_libev code review (docs/code-review-2026-06-10.md):

#4 tunnel_client.cpp data_from_local_or_fec_timeout: in port-range mode the
"handshake not complete" case early-returned before draining the local_listen
fd. libev's epoll is level-triggered, so an app packet arriving during the
handshake/re-handshake window spun the process at 100% CPU until HELLO_ACK
(forever if the server never came up). The same early return also swallowed
the one-shot FEC-timeout flush, stranding the half-filled encode group.
Fix: replace the early return with a can_send flag; fall through to drain the
socket and flush FEC as usual, and only gate the final send loop on can_send.
Verified: feeding packets to a serverless port-range client now holds 0% CPU.

#8 port_range_manager / pick_next_endpoint: rr_counter / reply_rr_counter were
signed ints incremented per send and never wrapped. After 2^31 sends (under a
day at high pps) signed overflow is UB; in port_range_manager the (int) cast
made `ports[rr_counter % size]` a negative OOB index. Switch both counters to
size_t (defined wraparound, never negative) and use bounded increment in
next_dest.

Verified: make (static) clean, tests/smoke.sh and tests/smoke_port_range.sh
both pass (100/100 delivery).

+++ connection.h
@@ -265,7 +265,7 @@ struct conn_info_t : not_copy_able_t  // stores info for a raw connection.for cl
         my_time_t last_seen_us;
     };
     std::vector<nat_endpoint_t> active_endpoints;
-    int reply_rr_counter = 0;
+    size_t reply_rr_counter = 0;   // unsigned: wraps cleanly, never goes negative
 
     void record_endpoint(const address_t &src, int fd_idx) {
         my_time_t now = get_current_time_us();
